src/requirements.txt
avalon_framework
bs4
pysnooper
python-telegram-bot
selenium
